fix: correct color order for tip and enable to change tip labels
The coloring by `fac` argument was incorrect because the order of the factors
returned by the `fac_dipatcher` function was different from the order of
the tips(leaves) in dendrograms, which has been corrected. Also, the code for
changing tip labels by `fac` and/or `labels` arguments was ineffective,
which has been corrected.